Fix frac-line
Render frac-line, underline, overline, and vertical-separator using an SVG image.

The trick is that the line consists of an SVG path that fills only the middle third of the SVG image and the containing span.

So if browser rounds down the size of the span, the entire line should still be visible.