dist script packs release tarball/zip with "katex" directory prefix again.
It does so by mercilessly removing the katex/ folder, if it exists
(presumably left by previous failed dist:zip attempts), copying dist/
to katex/, and archiving the latter. On success, katex/ gets removed.

Settled on this version after a more sophisticated script was considered
to use too much POSIX features (like set and trap) in #1665.

Deliberately not adding katex/ to .gitignore because if katex/ shows up,
it indicates that the latest dist attempt has not completed successfully.