Add --mode option to CLI
This option allows to set the output mode to either html, mathml or htmlAndMathml, as defined in the renderToString options.

Unfortunately, there is already an 'output' option that specifies the
file where the output should be written into.

I am aware that my solutions is most likely not idiomatic and looks
quite hacky. Since I am not a Javascript developer myself I would be
happy if someone can guide me to a proper implementation of this PR.

Closes #2237 .