fix: matrix environment with zero or inconsistent columns
Previously, the matrix family of environments assumed that the number of
columns was given by the number of columns in the first row.  This
didn't work when there were zero rows, and it didn't handle the case
where future rows have more columns than the first (which is visible
when using an alignment other than `c`, such as `\begin{matrix*}[r]`).

Fixes #3017