refactor: better `.pfa/.pfb` generation code
The old version was a big list of commands to `mftrace` in a single
target. Now the code uses a pattern to create the different files in a
more idiomatic way.

It is not necessary to explicitly state the encoding for some files.
The `mftrace` script can always find the correct encoding
automatically.