refactor: rewrite `makeFF` in Python
Rewriting `makeFF` in Python improves the code in several ways:
- Increases maintainability as Python code is more readable than
  dense Perl code using mostly string substitutions (sometimes with
  regular expressions) to write scripts, and rewrite other files.
- Pair kerning is a feature that requires us to read the `.tfm` files
  that contain those metric tables. The current codebase contains a
  `.tfm` parser written in Python, not in Perl. This commit makes it
  possible to use that to build pair kerning tables in the FontForge
  scripts.

This new code also splits the mapping between the old TeX font
codepoints and the modern Unicode scalar values into a separate module
that can in the future be reused for metrics generation. This decreases
code duplication, and makes changes more robust. It's also easier to
write sanity checks in the new mappings to catch easy-to-miss mistakes
like codepoints being mapped twice.