refactor: rewrite `mapping.pl` in Python
This rewrite is almost trivial, as we can reuse the mapping from the
font generation code. Because this deduplicates code, there is no need
to keep the mappings in sync. The only real rewrite is the inversion of
the mapping from the original Perl code, the rest of the code was a
duplication of the mapping.

This commit also contains the newly generated `fontMetricsData.js`.
The changes in this file correspond to adding all of the metrics of two
fonts (`Caligraphic-Bold` and `Fraktur-Bold`), and a few touch-ups for
several glyphs:
- In `Main-Italic` we add metrics for two glyphs:
  - 305 / 0x131: dotless i
  - 567 / 0x237: dotless j
- The rest of the additions are in `Typewriter-Regular`:
  - 168 / 0xA8: diereses
  - 710 / 0x2C6: circumflex
  - 729 / 0x2D9: dot above (but the glyph is a line below?)
  - 732 / 0x2DC: small tilde
  - 733 / 0x2DD: double acute accent

  729 and 733 are a little weird as the Unicode descriptions do not
  match their glyphs.
- Three glyphs are removed from `Typewriter-Regular`: (770, 771, 776)
  / (0x302, 0x303, 0x308). These are not present in the generated font
  files, so it's save to remove them.
- Only 126: 0x7E in `Typewriter-Regular` has changed metrics, its
  height and depth change.