feat(css): provide `katex-swap.css` that uses `font-display: swap` (#3940)
* Default to font-display: swap; closes #2242

Google fonts seem to default to swap, and mixing them with katex makes it so that katex is invisible briefly on page load, unlike google fonts which make use of fallbacks while loading. This would unify their behavior.

Ideally there'd be an option for this, and a ?display=swap in the cdn url (like google fonts: https://developers.google.com/fonts/docs/css2#use_font-display), but I don't know if jsdelivr supports that!

* Restore `block` default, provide `katex-swap.css` for `swap`

* Document katex-swap.css

* Fix dev mode

* Add missing katex-swap.scss

* Build scss separate from js, fix version

Use `webpack-remove-empty-scripts` instead of `splitChunks`

* lint fix

---------

Co-authored-by: Erik Demaine <edemaine@mit.edu>